// src/collections/HomePage/hooks/revalidatePage.ts
import type { HomePage } from '@/payload-types'
import type { CollectionAfterChangeHook, CollectionAfterDeleteHook } from 'payload'

import { revalidatePath, revalidateTag } from 'next/cache'

export const revalidatePage: CollectionAfterChangeHook<HomePage> = async ({
  doc,
  previousDoc,
  req: { payload, context },
}) => {
  if (!context.disableRevalidate) {
    const tenantDoc = await payload.findByID({
      collection: 'slug',
      // @ts-ignore
      id: doc.tenant,
      depth: 1,
    })

    if (doc._status === 'published') {
      // @ts-ignore
      const path = tenantDoc.name ? `/${tenantDoc.name}` : ''
      payload.logger.info(`Revalidating page at path: ${path}`)

      revalidatePath(path)
      revalidateTag('pages-sitemap', 'max')
    }

    // If the page was previously published, we need to revalidate the old path
    if (previousDoc?._status === 'published' && doc._status !== 'published') {
      // @ts-ignore
      const oldPath = tenantDoc.name ? `/${tenantDoc.name}` : ''

      payload.logger.info(`Revalidating old page at path: ${oldPath}`)

      revalidatePath(oldPath)
      revalidateTag('pages-sitemap', 'max')
    }
  }
  return doc
}

export const revalidateDelete: CollectionAfterDeleteHook<HomePage> = async ({
  doc,
  req: { context, payload },
}) => {
  if (!context.disableRevalidate) {
    const tenantDoc = await payload.findByID({
        collection: 'slug',
        // @ts-ignore
        id: doc.tenant,
        depth: 1,
      })

    // @ts-ignore
    const path = tenantDoc.name ? `/${tenantDoc.name}` : ''
    revalidatePath(path)
    revalidateTag('pages-sitemap', 'max')
  }

  return doc
}
